스토어저장

+++ src/main/java/com/metacoding/storev1/store/StoreService.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class StoreService {
 
+    public static final Store 상품수정 = null;
     public static Object 상세보기;
     private StoreRepository storeRepository;
 
@@ -20,7 +21,16 @@ public class StoreService {
     // 3번:board 프로젝트의 BoardService 참고
     @Transactional // insert,delete, update시에 사용 : 함수 종료시 commit 됨
     public void 상품삭제(int id) {
+
+        Store store = storeRepository.findById(id);
+
+        // 2. 삭제
+        if (store == null) {
+            throw new RuntimeException("삭제할 상품이 없습니다.");
+        }
+
         storeRepository.deleteById(id);
+
     }
 
     @Transactional
@@ -36,6 +46,21 @@ public class StoreService {
     public Store 상세보기(int id) {
         Store store = storeRepository.findById(id);
         return store;
+    }
+
+    @Transactional
+    public void 상품수정(int id, String name, int stock, int price) {
+        // 1.상품 조회
+        Store store = storeRepository.findById(id);
+
+        // 2. 없으면 터트리기
+        if (store == null) {
+            throw new RuntimeException("수정할 상품이 없습니다.");
+        }
+
+        // 3.상품수정
+        storeRepository.updateById(id, name, stock, price);
+
     }
 
 }
